+++ R/args.R
@@ -168,6 +168,16 @@ process_args <- function(args, app) {
     app$exprs[[spec$.val_pos_in_exprs]] <- val
   }
 
+  command_names <- setdiff(names(app_commands), ".val_pos_in_exprs")
+  if (
+    !length(positional_args) &&
+      length(command_names) &&
+      isTRUE(attr(app_commands, "help_on_missing_command"))
+  ) {
+    print_app_help(app, command_path = command_path, yaml = FALSE)
+    return(FALSE)
+  }
+
   if (length(positional_args) || length(app_args)) {
     # we've parsed all the command line args,
     # we can now match positional args